# SELLER_POSTMAN_OPENAPI.yaml
openapi: '3.0.0'
info:
  title: Ecom Seller API
  version: '2.1.0'
  description: >-
    API dành cho luồng đăng ký chờ duyệt và các chức năng quản trị Seller.
    Trừ register, application, application uploads và public shop, mọi endpoint
    /seller đều yêu cầu Users.role=seller và Sellers.status=active; nếu không BE trả
    403 với code SELLER_NOT_ACTIVE.
  contact:
    name: BacktoBach
servers:
  - url: http://localhost:5000/api
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /seller/register:
    post:
      tags: [Seller Application]
      summary: Gửi yêu cầu mở cửa hàng để Admin duyệt
      description: >-
        Tạo đơn ở trạng thái pending. API không đổi Users.role, không cấp seller JWT
        và cũng không trả accessToken mới. Đơn rejected được phép cập nhật và gửi lại.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SellerProfileInput'
      responses:
        '200':
          description: Đã tạo hoặc gửi lại đơn; tài khoản hiện tại vẫn giữ role customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellerApplicationSubmitResponse'
              example:
                status: success
                message: Yêu cầu mở cửa hàng đã được gửi và đang chờ duyệt.
                data:
                  application:
                    sellerId: sel_abc123
                    status: pending
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: Cửa hàng đang suspended, không được gửi lại đơn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: fail
                code: SELLER_SUSPENDED
                message: Cửa hàng đang bị tạm ngừng và không thể gửi lại đơn.
        '409':
          description: Đơn đang pending, shop đã active hoặc tên shop đã được dùng
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                pending:
                  value:
                    status: fail
                    code: SELLER_APPLICATION_PENDING
                    message: Yêu cầu mở cửa hàng đang chờ duyệt.
                active:
                  value:
                    status: fail
                    code: SELLER_ALREADY_ACTIVE
                    message: Cửa hàng đã được kích hoạt.
                duplicateShopName:
                  value:
                    status: fail
                    code: SHOP_NAME_TAKEN
                    message: Tên cửa hàng đã được sử dụng. Vui lòng chọn tên khác.
  /seller/application:
    get:
      tags: [Seller Application]
      summary: Lấy trạng thái đơn mở cửa hàng của tài khoản hiện tại
      description: >-
        Customer gọi endpoint này sau khi gửi đơn hoặc khi mở lại trang Become Seller.
        API không trả CCCD và thông tin ngân hàng. data.application là null nếu chưa từng gửi đơn.
      responses:
        '200':
          description: Trạng thái đơn hiện tại hoặc null
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellerApplicationStatusResponse'
              examples:
                hasApplication:
                  value:
                    status: success
                    data:
                      application:
                        sellerId: sel_abc123
                        shopName: Volitify Store
                        status: pending
                        createdAt: '2026-07-30T08:00:00.000Z'
                        updatedAt: '2026-07-30T08:00:00.000Z'
                noApplication:
                  value:
                    status: success
                    data:
                      application: null
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /seller/profile:
    get:
      tags: [Seller Profile]
      summary: Lấy hồ sơ shop hiện tại
      responses:
        '200':
          description: Hồ sơ shop
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Seller Profile]
      summary: Cập nhật hồ sơ shop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SellerProfileInput'
      responses:
        '200':
          description: Cập nhật hồ sơ thành công
        '400':
          $ref: '#/components/responses/BadRequest'
  /seller/shops/{shopId}:
    get:
      tags: [Seller Profile]
      summary: Xem thông tin shop công khai
      security: []
      parameters:
        - $ref: '#/components/parameters/ShopId'
      responses:
        '200':
          description: Shop, sản phẩm và thống kê công khai
        '404':
          $ref: '#/components/responses/NotFound'
...
  /seller/dashboard-analytics:
    get:
      tags: [Seller Dashboard]
      summary: Lấy dữ liệu biểu đồ theo ngày, tháng hoặc năm
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [day, month, year]
            default: day
        - $ref: '#/components/parameters/FromDate'
        - $ref: '#/components/parameters/ToDate'
      responses:
        '200':
          description: Summary, series và top products
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /seller/products:
    get:
      tags: [Seller Products]
      summary: Lấy sản phẩm thuộc shop hiện tại
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - name: status
          in: query
          schema:
            type: string
            enum: [all, active, inactive, low_stock, out_of_stock]
            default: all
        - name: categoryId
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [created_at, name, price, stock]
            default: created_at
        - $ref: '#/components/parameters/SortOrder'
      responses:
        '200':
          description: data.products và data.pagination
    post:
      tags: [Seller Products]
      summary: Tạo sản phẩm với một default variant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateInput'
      responses:
        '201':
          description: Tạo sản phẩm thành công
        '400':
          $ref: '#/components/responses/BadRequest'
  /seller/products/{productId}:
    put:
      tags: [Seller Products]
      summary: Cập nhật sản phẩm thuộc shop
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateInput'
      responses:
        '200':
          description: Cập nhật sản phẩm thành công
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [Seller Products]
      summary: Ẩn sản phẩm khỏi gian hàng
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Sản phẩm đã được ẩn
  /seller/inventory/low-stock:
    get:
      tags: [Seller Inventory]
      summary: Lấy sản phẩm sắp hết hoặc hết hàng
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: data.variants và data.pagination
  /seller/inventory/logs:
    get:
      tags: [Seller Inventory]
      summary: Lấy lịch sử thay đổi tồn kho
      parameters:
        - name: variantId
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [sale, order_cancelled, restock, manual_adjustment, return_refund]
        - $ref: '#/components/parameters/FromDate'
        - $ref: '#/components/parameters/ToDate'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: data.logs và data.pagination
  /seller/inventory/adjust:
    post:
      tags: [Seller Inventory]
      summary: Nhập kho hoặc điều chỉnh tồn kho thủ công
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryAdjustInput'
      responses:
        '200':
          description: Tồn kho và InventoryLog đã được cập nhật trong một transaction
        '409':
          description: Điều chỉnh làm tồn kho nhỏ hơn 0
  /seller/products/{productId}/variants/{variantId}/stock-alert:
    patch:
      tags: [Seller Inventory]
      summary: Cập nhật ngưỡng cảnh báo tồn kho
      parameters:
        - $ref: '#/components/parameters/ProductId'
        - $ref: '#/components/parameters/VariantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lowStockThreshold]
              properties:
                lowStockThreshold:
                  type: integer
                  minimum: 0
                  maximum: 1000000
                  example: 5
      responses:
        '200':
          description: Cập nhật ngưỡng thành công
  /seller/orders:
    get:
      tags: [Seller Orders]
      summary: Lấy đơn hàng có sản phẩm thuộc shop
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - name: status
          in: query
          schema:
            type: string
            enum: [all, pending_fulfillment, ready_to_ship, shipping, delivered, cancelled]
            default: all
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [created_at, total, status]
            default: created_at
        - $ref: '#/components/parameters/SortOrder'
      responses:
        '200':
          description: data.orders và data.pagination
  /seller/orders/{orderId}/timeline:
    get:
      tags: [Seller Orders]
      summary: Lấy timeline trạng thái đơn hàng thuộc shop
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Thông tin đơn, items và status history
        '404':
          $ref: '#/components/responses/NotFound'
  /seller/orders/items/{itemId}:
    patch:
      tags: [Seller Orders]
      summary: Cập nhật trạng thái một dòng đơn hàng
      description: 'Luồng hợp lệ: pending_fulfillment -> ready_to_ship -> shipping -> delivered. Có thể hủy trước shipping.'
      parameters:
        - $ref: '#/components/parameters/ItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderItemStatusInput'
      responses:
        '200':
          description: Cập nhật thành công và gửi notification cho customer
        '409':
          description: Chuyển trạng thái không hợp lệ
  /seller/coupons:
    get:
      tags: [Seller Coupons]
      summary: Lấy voucher thuộc shop
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - name: status
          in: query
          schema:
            type: string
            enum: [all, active, scheduled, expired, disabled, exhausted]
            default: all
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [created_at, code, starts_at, expires_at, used_count]
            default: created_at
        - $ref: '#/components/parameters/SortOrder'
      responses:
        '200':
          description: data.coupons và data.pagination
    post:
      tags: [Seller Coupons]
      summary: Tạo voucher cho shop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CouponCreateInput'
      responses:
        '201':
          description: Voucher được tạo
        '400':
          $ref: '#/components/responses/BadRequest'
  /seller/coupons/stats:
    get:
      tags: [Seller Coupons]
      summary: Lấy thống kê sử dụng voucher
      parameters:
        - $ref: '#/components/parameters/FromDate'
        - $ref: '#/components/parameters/ToDate'
        - name: status
          in: query
          schema:
            type: string
            enum: [all, active, scheduled, expired, disabled, exhausted]
            default: all
        - name: couponId
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Search'
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [created_at, code, redemptions, attributed_order_value, discount_amount, usage_rate]
            default: redemptions
        - $ref: '#/components/parameters/SortOrder'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Summary và thống kê từng voucher
  /seller/coupons/{couponId}:
    patch:
      tags: [Seller Coupons]
      summary: Cập nhật thời gian hoặc trạng thái voucher
      parameters:
        - $ref: '#/components/parameters/CouponId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CouponUpdateInput'
      responses:
        '200':
          description: Voucher được cập nhật
    delete:
      tags: [Seller Coupons]
      summary: Xóa mềm voucher và cho phép tạo lại code cũ
      parameters:
        - $ref: '#/components/parameters/CouponId'
      responses:
        '200':
          description: Voucher được xóa mềm
  /seller/flash-sales:
    get:
      tags: [Seller Flash Sales]
      summary: Lấy flash sale của shop
      responses:
        '200':
          description: data.flashSales
    post:
      tags: [Seller Flash Sales]
      summary: Tạo flash sale
      description: originalPrice phải gửi để tương thích validator hiện tại nhưng BE sẽ lấy lại giá gốc từ DB.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlashSaleCreateInput'
      responses:
        '201':
          description: Flash sale được tạo
        '400':
          $ref: '#/components/responses/BadRequest'
  /seller/flash-sales/{flashSaleId}:
    patch:
      tags: [Seller Flash Sales]
      summary: Cập nhật flash sale
      parameters:
        - $ref: '#/components/parameters/FlashSaleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlashSaleUpdateInput'
      responses:
        '200':
          description: Flash sale được cập nhật
    delete:
      tags: [Seller Flash Sales]
      summary: Ngừng flash sale bằng cách chuyển status thành inactive
      parameters:
        - $ref: '#/components/parameters/FlashSaleId'
      responses:
        '200':
          description: Flash sale đã ngừng
...
  /seller/returns:
    get:
      tags: [Seller Returns]
      summary: Lấy yêu cầu trả hàng thuộc shop
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - name: status
          in: query
          description: Status public dùng để lọc và hiển thị
          schema:
            type: string
            enum: [all, requested, approved, rejected, received, refunded]
            default: all
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [requested_at, status, product_name, customer_name]
            default: requested_at
        - $ref: '#/components/parameters/SortOrder'
      responses:
        '200':
          description: data.returns và data.pagination
  /seller/returns/{returnId}:
    get:
      tags: [Seller Returns]
      summary: Lấy chi tiết và lịch sử yêu cầu trả hàng
      parameters:
        - $ref: '#/components/parameters/ReturnId'
      responses:
        '200':
          description: data.return và data.history
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Seller Returns]
      summary: Chấp nhận, từ chối hoặc xác nhận đã nhận lại hàng
      description: Request dùng status nội bộ để tương thích FE; response trả status public approved hoặc received.
      parameters:
        - $ref: '#/components/parameters/ReturnId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnStatusInput'
      responses:
        '200':
          description: Cập nhật thành công; item_returned hoàn kho và ghi InventoryLog
        '409':
          description: Transition không hợp lệ hoặc seller cố tự đặt refunded
  /seller/finance/summary:
    get:
      tags: [Seller Finance]
      summary: Lấy tổng quan tài chính read-only
      parameters:
        - $ref: '#/components/parameters/FromDate'
        - $ref: '#/components/parameters/ToDate'
      responses:
        '200':
          description: Doanh thu đã giao, voucher, return và giá trị chờ xử lý
        '400':
          description: Khoảng ngày sai hoặc FINANCE_FUTURE_DATE_NOT_ALLOWED
  /seller/finance/transactions:
    get:
      tags: [Seller Finance]
      summary: Lấy giao dịch tài chính read-only
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/FromDate'
        - $ref: '#/components/parameters/ToDate'
        - name: type
          in: query
          schema:
            type: string
            enum: [all, sale, return]
            default: all
      responses:
        '200':
          description: data.transactions và data.pagination
        '400':
          description: Khoảng ngày sai hoặc FINANCE_FUTURE_DATE_NOT_ALLOWED
...
  /chat/history/{partnerId}:
    get:
      tags: [Chat]
      summary: Lấy lịch sử chat với một user
      parameters:
        - $ref: '#/components/parameters/PartnerId'
      responses:
        '200':
          description: Tin nhắn giữa current user và partnerId
  /chat/read/{partnerId}:
    post:
      tags: [Chat]
      summary: Đánh dấu tin nhắn từ partner đã đọc
      parameters:
        - $ref: '#/components/parameters/PartnerId'
      responses:
        '200':
          description: Số tin nhắn đã cập nhật
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Dán accessToken, Postman tự thêm tiền tố Bearer
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Search:
      name: search
      in: query
      schema:
        type: string
        maxLength: 100
    SortOrder:
      name: sortOrder
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
    FromDate:
      name: from
      in: query
      description: Phải gửi cùng to, định dạng YYYY-MM-DD
      schema:
        type: string
        format: date
        example: '2026-07-01'
    ToDate:
      name: to
      in: query
      description: Phải gửi cùng from và không được là ngày tương lai
      schema:
        type: string
        format: date
        example: '2026-07-29'
    ShopId:
      name: shopId
      in: path
      required: true
      schema:
        type: string
      example: sel_001
    ProductId:
      name: productId
      in: path
      required: true
      schema:
        type: string
      example: prod_001
    VariantId:
      name: variantId
      in: path
      required: true
      schema:
        type: string
      example: var_001
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: string
      example: order_001
...
  responses:
    BadRequest:
      description: Dữ liệu hoặc query không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Thiếu token hoặc thông tin đăng nhập không đúng
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Không có quyền với tài nguyên
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Không tìm thấy tài nguyên
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: fail
        code:
          type: string
          nullable: true
        message:
          type: string
...
    SellerProfileInput:
      type: object
      required: [shopName, shopPhone, shopAddress]
      properties:
        shopName:
          type: string
          example: Volitify Store
        shopPhone:
          type: string
          pattern: '^0\d{9}$'
          example: '0987654321'
        shopAddress:
          type: string
          example: 123 Nguyễn Văn Linh, Đà Nẵng
        description:
          type: string
        logoUrl:
          type: string
          format: uri
          nullable: true
        logoPublicId:
          type: string
          nullable: true
        coverUrl:
          type: string
          format: uri
          nullable: true
        coverPublicId:
          type: string
          nullable: true
        pickupAddress:
          type: string
          nullable: true
        identityName:
          type: string
          nullable: true
        identityNumber:
          type: string
          pattern: '^\d{12}$'
          nullable: true
          example: '012345678901'
        bankName:
          type: string
          nullable: true
        bankAccountNo:
          type: string
          pattern: '^\d{6,20}$'
          nullable: true
          example: '1234567890'
        bankAccountHolder:
          type: string
          nullable: true
...
    SellerApplicationSubmitResponse:
      type: object
      required: [status, message, data]
      properties:
        status:
          type: string
          enum: [success]
        message:
          type: string
        data:
          type: object
          required: [application]
          properties:
            application:
              type: object
              required: [sellerId, status]
              properties:
                sellerId:
                  type: string
                status:
                  type: string
                  enum: [pending]
    SellerApplicationStatusResponse:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: object
          required: [application]
          properties:
            application:
              allOf:
                - $ref: '#/components/schemas/SellerApplication'
              nullable: true
...